irq and i386 : reuse code and fix pic bug

+++ kernel/arch/i386/irq/include/irq.h
@@ -5,10 +5,10 @@
 
 void init_irq(void);
 
-void irq_map(void *handler,uint32_t irq_num);
-void irq_mask(uint32_t irq_num);
-void irq_eoi(uint32_t irq_num);
-void irq_generic_map(void *handler,uint32_t irq_num);
+void irq_map(void *handler,uintmax_t irq_num);
+void irq_mask(uintmax_t irq_num);
+void irq_eoi(uintmax_t irq_num);
+void irq_generic_map(void *handler,uintmax_t irq_num);
 
 #define PIC_APIC 0x01
 #define PIC_PIC  0x02
